Rework RPC Worker for less lock contention
Instead of holding the lock on the RPC queue the entire duration of the
procedure call, the lock will instead only be held for the duration of
obtaining the oldest message from the queue.

In addition, the worker thread will not sleep in between messages if the
queue is too full.